+++ NEWS.md
@@ -2,10 +2,12 @@ News
 =====
 
 # volcano3D 1.0.1
-###### 05/06/2020
+###### 08/06/2020
 
 * update vignette for CRAN (remove WebGL)
 * update default plotly parameters
+* Allow optional colour coding of labels in radial plots
+* remove ggplot warnings of NA in geom_path 
 
 
 # volcano3D 1.0.0
@@ -26,4 +28,4 @@ News
 # volcano3D 0.1.0
 ###### 27/04/2020
 
-* This is the initial release of volcano3D
+* This is the initial build of volcano3D