+++ NEWS.md
@@ -2,13 +2,14 @@ News
 =====
 
 # volcano3D 1.0.1
-###### 16/06/2020
+###### 26/06/2020
 
 * update vignette for CRAN (remove WebGL)
 * update default plotly parameters
 * Allow optional colour coding of labels in radial plots
 * remove ggplot warnings of NA in geom_path 
 * Allow colour of grids and axes to be changed
+* Convert 3D labels to annotations 
 
 
 # volcano3D 1.0.0